fix(settings): revert the shared spinner to a dialog route
My previous commit rewrote the shared progress spinner as an OverlayEntry to
dodge the raw-dialog ratchet. That was wrong: removing an OverlayEntry during
tree finalization trips navigator.dart's '!_debugLocked' assertion, which broke
five tests in nostr_settings_screen_test -- two of which exist specifically to
pin this behaviour ("does not crash when navigation closes progress overlay
first", "dismisses progress overlay when key deletion fails"). In CI's merged
isolate those failures then cascaded into unrelated camera and notifications
suites in the same shard.

The DialogRoute version was correct. `_navigator.removeRoute(route)` guarded on
`route.isActive` is navigation-aware in a way an overlay entry is not, and those
tests were already guarding it.

So keep the dialog route exactly where the ratchet already permits it, and drop
the two new files. The shared entry point is now a top-level
startAccountDeletionFlow() in nostr_settings_screen.dart, called by the Settings
hub and Support Center -- settings_screen.dart already imported that file for
the route path, so this adds no new dependency direction. The raw-dialog
baseline is byte-identical to main again.

Refs #6335
