+++ mobile/lib/screens/settings/nostr_settings_screen.dart
@@ -7,6 +7,7 @@ import 'package:divine_ui/divine_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:models/models.dart';
 import 'package:nostr_client/nostr_client.dart' show Nip89ClientTag;
 import 'package:nostr_key_manager/nostr_key_manager.dart'
     show SecureKeyStorageException;
@@ -15,6 +16,8 @@ import 'package:openvine/features/feature_flags/providers/feature_flag_providers
 import 'package:openvine/l10n/l10n.dart';
 import 'package:openvine/providers/app_providers.dart';
 import 'package:openvine/providers/nostr_client_provider.dart';
+import 'package:openvine/providers/owned_divine_username_provider.dart';
+import 'package:openvine/providers/user_profile_providers.dart';
 import 'package:openvine/screens/auth/welcome_screen.dart';
 import 'package:openvine/screens/blossom_settings_screen.dart';
 import 'package:openvine/screens/key_management_screen.dart';
@@ -23,9 +26,8 @@ import 'package:openvine/screens/relay_settings_screen.dart';
 import 'package:openvine/screens/settings/nip05_settings_screen.dart';
 import 'package:openvine/services/auth_service.dart' hide UserProfile;
 import 'package:openvine/services/nostr_signature_verification_preference_service.dart';
-import 'package:openvine/widgets/delete_account_action.dart';
+import 'package:openvine/widgets/delete_account_confirmation.dart';
 import 'package:openvine/widgets/delete_account_dialog.dart';
-import 'package:openvine/widgets/modal_progress_overlay.dart';
 
 class NostrSettingsScreen extends ConsumerWidget {
   static const routeName = 'nostr-settings';
@@ -150,7 +152,9 @@ class _RemoveKeysTile extends StatelessWidget {
       onConfirm: () async {
         if (!context.mounted) return;
 
-        final progressOverlay = ModalProgressOverlay.show(context);
+        final progressOverlay = _ProgressOverlay.show(context);
+
+        unawaited(progressOverlay.closed);
 
         try {
           await authService.signOut(
@@ -190,6 +194,42 @@ class _RemoveKeysTile extends StatelessWidget {
     );
   }
 }
+
+class _ProgressOverlay {
+  _ProgressOverlay._(
+    this._navigator, {
+    required this.route,
+    required this.closed,
+  });
+
+  final NavigatorState _navigator;
+  final Route<void> route;
+  final Future<void> closed;
+  bool _dismissed = false;
+
+  static _ProgressOverlay show(BuildContext context) {
+    final navigator = Navigator.of(context, rootNavigator: true);
+    late final DialogRoute<void> route;
+    route = DialogRoute<void>(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) => const Center(
+        child: CircularProgressIndicator(color: VineTheme.vineGreen),
+      ),
+    );
+
+    final closed = navigator.push(route);
+    return _ProgressOverlay._(navigator, route: route, closed: closed);
+  }
+
+  void dismiss() {
+    if (_dismissed || !_navigator.mounted || !route.isActive) return;
+    _dismissed = true;
+    _navigator.removeRoute(route);
+  }
+}
+
+const Duration _profileResolveTimeout = Duration(seconds: 3);
 
 class _DeleteAccountTile extends StatelessWidget {
   const _DeleteAccountTile({required this.ref});
@@ -445,4 +485,84 @@ class _SectionHeader extends StatelessWidget {
       ),
     );
   }
+}
+
+/// Open the account-deletion confirmation flow.
+///
+/// The single implementation shared by every surface that offers deletion —
+/// the Settings hub, Support Center, and the Danger Zone below — so the
+/// confirmation gate, the username-burn opt-in, and the irreversible-step
+/// ordering cannot drift between entry points. [screenName] is used for log
+/// attribution only.
+///
+/// Lives here rather than under `widgets/` because the progress overlay it uses
+/// is a dialog route, and the raw-dialog ratchet only permits those in files
+/// that already have them.
+///
+/// Does nothing when no account is signed in.
+Future<void> startAccountDeletionFlow({
+  required BuildContext context,
+  required WidgetRef ref,
+  required String screenName,
+}) async {
+  final deletionService = ref.read(accountDeletionServiceProvider);
+  final authService = ref.read(authServiceProvider);
+  final profileRepository = ref.read(profileRepositoryProvider);
+  final pubkey = authService.currentPublicKeyHex;
+  if (pubkey == null || pubkey.isEmpty) return;
+
+  // Kick off the burnable-handle lookup but do not await it: the dialog opens
+  // immediately and reveals the opt-in burn toggle once this resolves, so a
+  // slow name-server call never blocks the tap.
+  final ownedUsernameFuture = ref.read(ownedDivineUsernameProvider.future);
+
+  // Resolve the local profile up front so the identity + username gate are
+  // ready when the dialog opens. Cache-first (usually instant) with a bounded
+  // network fallback; a timeout or miss degrades to npub + DELETE, never
+  // hangs.
+  final overlay = _ProgressOverlay.show(context);
+  UserProfile? profile;
+  try {
+    profile = await ref
+        .read(fetchUserProfileProvider(pubkey).future)
+        .timeout(_profileResolveTimeout, onTimeout: () => null);
+  } catch (_) {
+    profile = null;
+  } finally {
+    overlay.dismiss();
+  }
+  if (!context.mounted) return;
+
+  // The gate anchors on the shown profile identity (displayNip05) to confirm
+  // *which account* is being erased — deliberately distinct from the burn
+  // target (the owned @divine.video handle, which the burn toggle names for
+  // itself). The two can differ for an external-NIP-05 user who also owns a
+  // divine username; that divergence is intended, not a mismatch to reconcile.
+  final confirmation = DeleteAccountConfirmation(
+    pubkeyHex: pubkey,
+    displayName:
+        profile?.bestDisplayName ?? UserProfile.defaultDisplayNameFor(pubkey),
+    avatarUrl: profile?.picture,
+    handle: profile?.displayNip05,
+  );
+
+  await showDeleteAllContentWarningDialog(
+    context: context,
+    confirmation: confirmation,
+    ownedUsernameFuture: ownedUsernameFuture,
+    onConfirm:
+        ({
+          required bool burnUsername,
+          ({String name, String canonical})? ownedUsername,
+        }) => executeAccountDeletion(
+          context: context,
+          deletionService: deletionService,
+          authService: authService,
+          profileRepository: profileRepository,
+          burnUsername: burnUsername,
+          ownedUsername: ownedUsername,
+          confirmedPubkey: pubkey,
+          screenName: screenName,
+        ),
+  );
 }
